-- Scoped every batch reset to each unit's own job identifiers, so dispatching a batch no longer clears the succeeded record of a sibling unit that tracks the same stage under the same identifier, and widened the execution host reset operation to carry per-unit identifiers without costing a remote batch any extra round trips.
-- Fixed remote batch reporting end to end, so a settled batch reads what its jobs recorded rather than counting every one of them outstanding, every prepared batch a submission dispatched receives an outcome, a re-submitted batch keeps the allocations its first attempt queued, and the four terminal scheduler states the enumeration omitted now settle instead of being adopted forever.
-- Carried each job's recorded status and executor from the state artifact preparation already reads onto its descriptor, so remote reconciliation resolves a running allocation from the batch itself rather than from a tracker it would have to open across the transport while the batch runs.
-- Corrected the persisted acquisition data the assemblers write, covering the gas puff counter that recorded every delivered puff as zero, the runtime distance streams that placed trial boundaries and trigger zones a fixed offset away from the cue column, the trial distance that accumulated at a narrower width than it is stored at, and the ScanImage fallback that mis-ordered a multi-acquisition session.
-- Standardized the animal identifier as text across the manifest, job, dataset state, and plan artifacts, so a reader joins any pair of them without casting either side, and added a shared natural ordering so animal 2 precedes animal 10 in every table.
-- Pinned the timestamp, elapsed-minute, fluorescence, and microcontroller state column widths, so one dataset carries a single dtype per column whichever path assembled each of its sessions.
-- Made dataset definition recoverable and tolerant of partial coverage, so a definition interrupted before it recorded its cross-recording plans completes on the next identical call while a dataset whose earlier animals have moved to storage keeps growing, which is what lets a large project be forged in passes.
-- Released the resources the failure paths leaked or destroyed, covering the SSH clients an exhausted retry loop abandoned, the OpenMP runtime a link pointing at its own path replaced, and the camera timestamps a colliding manifest name overwrote.
-- Fixed the remaining reporting and refusal paths, covering the manifest published ahead of its job artifact, the checksum run refused before it overwrites a recorded verdict, the masked password placeholder refused before it replaces the stored one, the host budgets reported only for the machine they describe, and a pose-tracking estimate that dropped the whole video pipeline from a session's plan.
-- Added the missing test oracles, pinning the dataset admission set, the checksum exclusions, four serialization widths, the adopted allocation map, the state mirroring order, and the dataset arm of batch closure.

+++ src/sollertia_forgery/mesoscope_vr/microcontrollers.py
@@ -516,7 +516,7 @@ def _parse_ttl_data(event_partition: dict[int, pl.DataFrame], output_file: Path,
 
     if triggers[-1] != 0:
         timestamps = np.append(timestamps, values=timestamps[-1] + 1)
-        triggers = np.append(triggers, values=0)
+        triggers = np.append(triggers, values=np.uint8(0))
 
     result_dataframe = pl.DataFrame({"time_us": timestamps, "ttl_state": triggers})
     result_dataframe.write_ipc(file=output_file, compression="uncompressed")
@@ -618,7 +618,7 @@ def _parse_valve_data(
 
     if tone_states[-1] != 0:
         tone_timestamps = np.append(tone_timestamps, values=tone_timestamps[-1] + 1)
-        tone_states = np.append(tone_states, values=0)
+        tone_states = np.append(tone_states, values=np.uint8(0))
 
     shared_timestamps = np.unique(np.concatenate([tone_timestamps, reward_timestamps]))
 
@@ -682,9 +682,10 @@ def _parse_gas_puff_data(
         values_b=np.zeros(len(closed_timestamps), dtype=np.uint8),
     )
 
-    edges = np.diff(states, prepend=states[0])
+    # Differenced in a signed width, since an unsigned difference wraps a falling edge to 255 and never equals -1.
+    edges = np.diff(states.astype(np.int16), prepend=np.int16(states[0]))
     falling_edges = edges == -1
-    cumulative_puffs: NDArray[np.uint32] = np.cumsum(falling_edges.astype(np.uint32))
+    cumulative_puffs: NDArray[np.uint32] = np.cumsum(falling_edges, dtype=np.uint32)
 
     result_dataframe = pl.DataFrame(
         {"time_us": timestamps, "puff_state": states, "cumulative_puff_count": cumulative_puffs}
